requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
Babel==2.6.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
dj-database-url==0.5.0
Django==2.2
django-extensions==2.1.6
django-filter==2.1.0
django-phonenumber-field==1.1.0
django-rest-knox==4.0.1
djangorestframework==3.9.3
gunicorn==19.9.0
idna==2.8
isort==4.3.17
lazy-object-proxy==1.3.1
Markdown==3.1
mccabe==0.6.1
phonenumberslite==8.10.11
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.1
PySocks==1.6.8
pytz==2019.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
twilio==6.26.3
typed-ast==1.3.4
urllib3==1.24.3
whitenoise==4.1.2
wrapt==1.11.1
